+++ R/generateFromDate.R
@@ -6,7 +6,7 @@
 #' Taking Date or POSIXct colums, and building factor columns from them. 
 #' @param dataSet Matrix, data.frame or data.table
 #' @param cols list of date column(s) name(s) of dataSet to transform into factor. To transform all 
-#' dates, set it to "auto"
+#' dates, set it to "auto", (characters, default to "auto")
 #' @param type "year", "yearquarter", "yearmonth", "quarter" or "month", way to aggregate a date, 
 #' (character, default to "yearmonth")
 #' @param drop should \code{cols} be dropped after generation (logical, default to FALSE)
@@ -29,41 +29,33 @@
 #' head(messy_adult[, .(date1.quarter, date2.quarter)])
 #' @export
 #' @import data.table
-## Working environement
 generateFactorFromDate <- function(dataSet, cols, type = "yearmonth", drop = FALSE, verbose = TRUE, ...){
+  ## Working environement
   function_name <- "generateFactorFromDate"
   
   ## Sanity check
   dataSet <- checkAndReturnDataTable(dataSet)
-  cols <- real_cols(cols, names(dataSet), function_name = function_name)
+  cols <- real_cols(dataSet, cols, function_name, types = "date")
   is.verbose(verbose)
-  if (all(cols == "auto")){
-    cols = names(dataSet)[sapply(dataSet, is.date)]  
-  }
+  
   ## Initialization
   args <- list(...)
   name_separator <- build_name_separator(args)
   start_time <- proc.time()
-  n_transformed <- 0
+  
   ## Computation
   if (verbose){ 
     printl(function_name, ": I will create a factor column from each date column.")
     pb <- initPB(function_name, cols)
   }
   for (col in cols){
-    if (is.date(dataSet[[col]])){
-      new_col <- paste0(col, name_separator, type)
-      new_col <- make_new_col_name(new_col, names(dataSet))
-      dataSet[, (new_col) := date_factor(dataSet[[col]], type = type)]
-      n_transformed <- n_transformed + 1
-      if (isTRUE(drop)){
-        dataSet[, c(col) := NULL]
-      }
+    new_col <- paste0(col, name_separator, type)
+    new_col <- make_new_col_name(new_col, names(dataSet))
+    dataSet[, (new_col) := date_factor(dataSet[[col]], type = type)]
+    if (isTRUE(drop)){
+      dataSet[, c(col) := NULL]
     }
-    else{
-      warning(paste0(function_name, ": ", col, " isn't a date, i do nothing."))
-      
-    }
+  
     if (verbose){
       setPB(pb, col)
     }
@@ -71,7 +63,7 @@ generateFactorFromDate <- function(dataSet, cols, type = "yearmonth", drop = FAL
   if (verbose){ 
 	gc(verbose = FALSE)
     printl(function_name, ": It took me ", round( (proc.time() - start_time)[[3]], 2), 
-           "s to transform ", n_transformed, " column(s).")
+           "s to transform ", length(cols), " column(s).")
   }
   
   ## Wrapp-up
@@ -182,18 +174,8 @@ generateDateDiffs <- function(dataSet, cols, analysisDate = NULL, units = "years
   if (!is.null(analysisDate) & ! is.date(analysisDate)){
     stop(paste0(function_name, ": analysisDate must be a Date"))
   }
-  if (all(cols == "auto")){
-    cols = names(dataSet)[sapply(dataSet, is.date)]  
-  }
-  else{
-    cols <- real_cols(cols, names(dataSet), function_name = function_name)
-    for (col in cols){
-      if (! is.date(dataSet[[col]])){
-        warning(paste0(function_name, ": ", col, " isn't a date column, i don't do anything with it."))
-        cols <- cols[col != cols]
-      }
-    }
-  }
+  cols <- real_cols(dataSet, cols, function_name, types = "date")
+
   ## Initialization
   args <- list(...)
   name_separator <- build_name_separator(args)